requirements.txt
spotipy>=2.24.0
python-dotenv>=1.0.0
numpy>=1.26.0
requests>=2.31.0
anthropic>=0.40.0
pydantic>=2.0.0
